+++ test/Test.hs
@@ -1,7 +1,9 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE QuasiQuotes #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TemplateHaskell #-}
@@ -21,6 +23,12 @@ import Data.Aeson.GADT.TH
 import Expectations
 import Test.Hspec
 
+#if MIN_VERSION_dependent_sum(0,5,0)
+#else
+pattern Some :: tag a -> Some tag
+pattern Some x = This x
+#endif
+
 main :: IO ()
 main = hspec $ do
   describe "aeson-gadt-th" $ do
@@ -29,9 +37,9 @@ main = hspec $ do
       toJSON (Baz 1.2) `shouldBe` [aesonQQ| ["Baz", 1.2] |]
     it "should generate an expected FromJSON Some instance" $ do
       fromJSON [aesonQQ| ["Bar", "a"] |]
-        `shouldMatchPattern_` (\case Success (This (Bar 'a')) -> ())
+        `shouldMatchPattern_` (\case Success (Some (Bar 'a')) -> ())
       fromJSON [aesonQQ| ["Baz", 1.2] |]
-        `shouldMatchPattern_` (\case Success (This (Baz 1.2)) -> ())
+        `shouldMatchPattern_` (\case Success (Some (Baz 1.2)) -> ())
       (fromJSON [aesonQQ| ["bad", "input"] |] :: Result (Some Foo))
         `shouldMatchPattern_` (\case Error "Expected tag to be one of [Bar, Baz] but got: bad" -> ())
 
@@ -40,9 +48,9 @@ main = hspec $ do
       toJSON (Spam'Life 1.2) `shouldBe` [aesonQQ| ["Life", 1.2] |]
     it "should generate an expected FromJSON Some instance with constructor modifier" $ do
       fromJSON [aesonQQ| ["Eggs", "a"] |]
-        `shouldMatchPattern_` (\case Success (This (Spam'Eggs 'a')) -> ())
+        `shouldMatchPattern_` (\case Success (Some (Spam'Eggs 'a')) -> ())
       fromJSON [aesonQQ| ["Life", 1.2] |]
-        `shouldMatchPattern_` (\case Success (This (Spam'Life 1.2)) -> ())
+        `shouldMatchPattern_` (\case Success (Some (Spam'Life 1.2)) -> ())
       (fromJSON [aesonQQ| ["bad", "input"] |] :: Result (Some Spam))
         `shouldMatchPattern_` (\case Error "Expected tag to be one of [Eggs, Life] but got: bad" -> ())
 
@@ -51,9 +59,9 @@ main = hspec $ do
       toJSON (Gamma 1.2) `shouldBe` [aesonQQ| {"type": "Gamma", "value": 1.2} |]
     it "should generate an expected FromJSON Some instance with tagged object" $ do
       fromJSON [aesonQQ| {"type": "Beta", "value": "a"} |]
-        `shouldMatchPattern_` (\case Success (This (Beta 'a')) -> ())
+        `shouldMatchPattern_` (\case Success (Some (Beta 'a')) -> ())
       fromJSON [aesonQQ| {"type": "Gamma", "value": 1.2} |]
-        `shouldMatchPattern_` (\case Success (This (Gamma 1.2)) -> ())
+        `shouldMatchPattern_` (\case Success (Some (Gamma 1.2)) -> ())
       (fromJSON [aesonQQ| {"type": "bad", "value": "input"} |] :: Result (Some Alpha))
         `shouldMatchPattern_` (\case Error "Expected tag to be one of [Beta, Gamma] but got: bad" -> ())